Reorganize src/test
