+++ src/test/java/ch/psi/ord/core/RoCrateExporterTest.java
@@ -1,4 +1,4 @@
-package ch.psi.scicat;
+package ch.psi.ord.core;
 
 import io.quarkus.test.junit.QuarkusTest;
 import jakarta.inject.Inject;
